move Chef::VersionString to Chef::Utils
